Make Int a word-size type.
- Int and UInt are now struct types backed by Builtin.Word. Previously they
were typealiases for Int64; Int and Int64 are now distinct types.
- Mangled names 'i' and 'u' are now Int and UInt. Int64 is mangled longhand.
- Word is a typealias for Int. It is expected to go away in the future.
- Builtin.Word is unchanged.
- CLong and CUnsignedLong are typealiases for Int and UInt.
- FixedPoint.swift is now FixedPoint32.swift and FixedPoint64.swift. 
Reunifying these requires better builtins, especially for checked 
conversions (rdar://15472770).
- Updated many tests, mostly because Int is no longer spelled Int64 in sil.
- One check was removed from test decl/operator/operators.swift 
because it changed behavior when Int became a non-typealias 
type (rdar://15934688).



Swift SVN r13109

+++ lib/AST/Mangle.cpp
@@ -1028,10 +1028,10 @@ bool Mangler::tryMangleStandardSubstitution(NominalTypeDecl *decl) {
   // reserved for the swift module itself) or a digit or '_'.
 
   StringRef name = decl->getName().str();
-  if (name == "Int64") {
+  if (name == "Int") {
     Buffer << "Si";
     return true;
-  } else if (name == "UInt64") {
+  } else if (name == "UInt") {
     Buffer << "Su";
     return true;
   } else if (name == "Bool") {
